LPD-8428 Rename to pageEditorPagesTest
