+++ modules/test/playwright/tests/layout-content-page-editor-web/undoRedo.spec.ts
@@ -10,7 +10,7 @@ import {applicationsMenuPageTest} from '../../fixtures/applicationsMenuPageTest'
 import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
 import {loginTest} from '../../fixtures/loginTest';
 import getRandomId from '../../utils/getRandomId';
-import {pageEditorPagesTest} from './fixtures/pageEditorPages';
+import {pageEditorPagesTest} from './fixtures/pageEditorPagesTest';
 import getFragmentDefinition from './utils/getFragmentDefinition';
 import getPageDefinition from './utils/getPageDefinition';
